refactor: relocate interfaces from implementation to consumer packages
Move interface definitions to the packages that consume them per Go best
practices. Each cmd/ package now owns its interface (MailClient,
CalendarClient, ContactsClient, DriveClient) and its mock. This removes
the coupling between implementation and consumer packages and eliminates
the centralized testutil/mocks.go file.
