+++ internal/auth/auth_test.go
@@ -1,9 +1,9 @@
 package auth
 
 import (
-	"strings"
 	"os"
 	"path/filepath"
+	"strings"
 	"testing"
 
 	"github.com/open-cli-collective/google-readonly/internal/config"
@@ -17,17 +17,17 @@ func TestDeprecatedWrappers(t *testing.T) {
 
 		authDir, err := GetConfigDir()
 		if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
+			t.Fatalf("unexpected error: %v", err)
+		}
 
 		configDir, err := config.GetConfigDir()
 		if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
+			t.Fatalf("unexpected error: %v", err)
+		}
 
 		if authDir != configDir {
-		t.Errorf("got %v, want %v", authDir, configDir)
-	}
+			t.Errorf("got %v, want %v", authDir, configDir)
+		}
 	})
 
 	t.Run("GetCredentialsPath delegates to config package", func(t *testing.T) {
@@ -36,17 +36,17 @@ func TestDeprecatedWrappers(t *testing.T) {
 
 		authPath, err := GetCredentialsPath()
 		if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
+			t.Fatalf("unexpected error: %v", err)
+		}
 
 		configPath, err := config.GetCredentialsPath()
 		if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
+			t.Fatalf("unexpected error: %v", err)
+		}
 
 		if authPath != configPath {
-		t.Errorf("got %v, want %v", authPath, configPath)
-	}
+			t.Errorf("got %v, want %v", authPath, configPath)
+		}
 	})
 
 	t.Run("GetTokenPath delegates to config package", func(t *testing.T) {
@@ -55,24 +55,24 @@ func TestDeprecatedWrappers(t *testing.T) {
 
 		authPath, err := GetTokenPath()
 		if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
+			t.Fatalf("unexpected error: %v", err)
+		}
 
 		configPath, err := config.GetTokenPath()
 		if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
+			t.Fatalf("unexpected error: %v", err)
+		}
 
 		if authPath != configPath {
-		t.Errorf("got %v, want %v", authPath, configPath)
-	}
+			t.Errorf("got %v, want %v", authPath, configPath)
+		}
 	})
 
 	t.Run("ShortenPath delegates to config package", func(t *testing.T) {
 		home, err := os.UserHomeDir()
 		if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
+			t.Fatalf("unexpected error: %v", err)
+		}
 
 		testPath := filepath.Join(home, ".config", "test")
 
@@ -80,20 +80,20 @@ func TestDeprecatedWrappers(t *testing.T) {
 		configResult := config.ShortenPath(testPath)
 
 		if authResult != configResult {
-		t.Errorf("got %v, want %v", authResult, configResult)
-	}
+			t.Errorf("got %v, want %v", authResult, configResult)
+		}
 	})
 
 	t.Run("Constants match config package", func(t *testing.T) {
 		if ConfigDirName != config.DirName {
-		t.Errorf("got %v, want %v", ConfigDirName, config.DirName)
-	}
+			t.Errorf("got %v, want %v", ConfigDirName, config.DirName)
+		}
 		if CredentialsFile != config.CredentialsFile {
-		t.Errorf("got %v, want %v", CredentialsFile, config.CredentialsFile)
-	}
+			t.Errorf("got %v, want %v", CredentialsFile, config.CredentialsFile)
+		}
 		if TokenFile != config.TokenFile {
-		t.Errorf("got %v, want %v", TokenFile, config.TokenFile)
-	}
+			t.Errorf("got %v, want %v", TokenFile, config.TokenFile)
+		}
 	})
 }
 
@@ -129,29 +129,29 @@ func TestTokenFromFile(t *testing.T) {
 		}`
 		err := os.WriteFile(tokenPath, []byte(tokenData), 0600)
 		if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
+			t.Fatalf("unexpected error: %v", err)
+		}
 
 		token, err := tokenFromFile(tokenPath)
 		if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
+			t.Fatalf("unexpected error: %v", err)
+		}
 		if token.AccessToken != "test-access-token" {
-		t.Errorf("got %v, want %v", token.AccessToken, "test-access-token")
-	}
+			t.Errorf("got %v, want %v", token.AccessToken, "test-access-token")
+		}
 		if token.TokenType != "Bearer" {
-		t.Errorf("got %v, want %v", token.TokenType, "Bearer")
-	}
+			t.Errorf("got %v, want %v", token.TokenType, "Bearer")
+		}
 		if token.RefreshToken != "test-refresh-token" {
-		t.Errorf("got %v, want %v", token.RefreshToken, "test-refresh-token")
-	}
+			t.Errorf("got %v, want %v", token.RefreshToken, "test-refresh-token")
+		}
 	})
 
 	t.Run("returns error for non-existent file", func(t *testing.T) {
 		_, err := tokenFromFile("/nonexistent/token.json")
 		if err == nil {
-		t.Fatal("expected error, got nil")
-	}
+			t.Fatal("expected error, got nil")
+		}
 	})
 
 	t.Run("returns error for invalid JSON", func(t *testing.T) {
@@ -160,12 +160,12 @@ func TestTokenFromFile(t *testing.T) {
 
 		err := os.WriteFile(tokenPath, []byte("not valid json"), 0600)
 		if err != nil {
-		t.Fatalf("unexpected error: %v", err)
-	}
+			t.Fatalf("unexpected error: %v", err)
+		}
 
 		_, err = tokenFromFile(tokenPath)
 		if err == nil {
-		t.Fatal("expected error, got nil")
-	}
+			t.Fatal("expected error, got nil")
+		}
 	})
 }
