+++ internal/cmd/mail/attachments_download.go
@@ -135,7 +135,7 @@ Examples:
 	return cmd
 }
 
-func downloadAttachment(client gmail.GmailClientInterface, messageID string, att *gmail.Attachment) ([]byte, error) {
+func downloadAttachment(client MailClient, messageID string, att *gmail.Attachment) ([]byte, error) {
 	if att.AttachmentID != "" {
 		return client.DownloadAttachment(messageID, att.AttachmentID)
 	}
